:tada: Initialize auth flag; Allow for flags that are not features
