+++ .env.example
@@ -62,6 +62,7 @@ SUBSCRIPTION_KEY_GENERATOR_SECRET=
 OPENSSL_CONF=
 
 # Feature flags
+AUTHENTICATION_FALLBACK=
 FEATURE_EMAIL_ALERTS=
 FEATURE_READ_ONLY=
 FEATURE_SIGN_IN_ALERT=
